LayoutTests/storage/domstorage/localstorage/blocked-file-access-expected.txt

CONSOLE MESSAGE: line 11: PASS: window.localStorage is NOT accessible
CONSOLE MESSAGE: line 12: Exception: The operation is insecure.

Test that we cannot access localStorage from a file URL if unversal access is turned off.
